+++ test/Unit Tests/140) Snapshot Tests.js
@@ -3,7 +3,6 @@
 const assert = require( 'assert' );
 const jsongin = require( '../../src/jsongin' )
 	.NewJsongin( {
-		PathExtensions: false,
 		Explain: false,
 	} );